creted controller class and service class

+++ src/main/java/com/example/repository/EmployeeRepository.java
@@ -4,6 +4,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.entity.Employee;
 
-public interface EmployeeRepository extends JpaRepository<Employee,Integer> {
+public interface EmployeeRepository extends JpaRepository<Employee,Long> {
+
+	 
+	 
 
 }
